yatube/requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
django-debug-toolbar==3.2.4
Faker==12.0.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
mixer==7.1.2
packaging==21.3
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
python-dotenv~=0.21.0